Rename Store to ForkChoice.
