+++ lib/lambda_ethereum_consensus/beacon/sync_blocks.ex
@@ -9,7 +9,7 @@ defmodule LambdaEthereumConsensus.Beacon.SyncBlocks do
 
   alias LambdaEthereumConsensus.Beacon.BeaconChain
   alias LambdaEthereumConsensus.Beacon.PendingBlocks
-  alias LambdaEthereumConsensus.ForkChoice.Store
+  alias LambdaEthereumConsensus.ForkChoice
   alias LambdaEthereumConsensus.P2P.BlockDownloader
   alias LambdaEthereumConsensus.StateTransition.Misc
 
@@ -24,7 +24,7 @@ defmodule LambdaEthereumConsensus.Beacon.SyncBlocks do
   def run(_opts) do
     # Initial sleep for faster app start
     Process.sleep(1000)
-    {:ok, checkpoint} = Store.get_finalized_checkpoint()
+    {:ok, checkpoint} = ForkChoice.get_finalized_checkpoint()
 
     initial_slot = Misc.compute_start_slot_at_epoch(checkpoint.epoch)
     last_slot = BeaconChain.get_current_slot()
