+++ lib/lambda_ethereum_consensus/fork_choice/fork_choice.ex
@@ -1,4 +1,4 @@
-defmodule LambdaEthereumConsensus.ForkChoice.Store do
+defmodule LambdaEthereumConsensus.ForkChoice do
   @moduledoc """
     The Store is responsible for tracking information required for the fork choice algorithm.
   """
@@ -29,24 +29,6 @@ defmodule LambdaEthereumConsensus.ForkChoice.Store do
   def get_finalized_checkpoint do
     [finalized_checkpoint] = get_store_attrs([:finalized_checkpoint])
     {:ok, finalized_checkpoint}
-  end
-
-  @spec get_current_status_message() :: {:ok, Types.StatusMessage.t()} | {:error, any}
-  def get_current_status_message do
-    # TODO: un-hardcode when get_head is optimized and/or cached
-    # GenServer.call(__MODULE__, :get_current_status_message, @default_timeout)
-
-    # hardcoded response from random peer
-    {:ok,
-     %Types.StatusMessage{
-       fork_digest: Base.decode16!("BBA4DA96"),
-       finalized_root:
-         Base.decode16!("7715794499C07D9954DD223EC2C6B846D3BAB27956D093000FADC1B8219F74D4"),
-       finalized_epoch: 228_168,
-       head_root:
-         Base.decode16!("D62A74AE0F933224133C5E6E1827A2835A1E705F0CDFEE3AD25808DDEA5572DB"),
-       head_slot: 7_301_450
-     }}
   end
 
   @spec has_block?(Types.root()) :: boolean()
